pyproject.toml
[tool.poetry]
name = "aws-cloudwatch-docker"
version = "0.1.0"
description = ""
authors = ["Nikolai Zinchenko <nikolaizinchenko94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
docker = "^7.0.0"
boto3 = "^1.34.31"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
